Add syncfusion_flutter_gauges dependency, update PrestamoChofer to use double for mileage, and enhance SolicitudChoferRepository with new methods

+++ lib/domain/driver-car/PrestamoChofer.dart
@@ -15,8 +15,8 @@ class PrestamoChofer {
     int? codSucursal;
     DateTime? fechaEntrega;
     int? codEmpEntregadoPor;
-    int? kilometrajeEntrega;
-    int? kilometrajeRecepcion;
+    double? kilometrajeEntrega;
+    double? kilometrajeRecepcion;
     int? nivelCombustibleEntrega;
     int? nivelCombustibleRecepcion;
     int? estadoLateralesEntrega;
